refactor: split desktop test mock helpers
